\name{is_rookable}
\alias{is_rookable}
\title{
Test for Rookable applications
}
\description{
A convenience function for testing whether or not objects are either a function or reference class as defined by the Rook specification for applications.
}
\usage{
is_rookable(app)
}
\arguments{
  \item{app}{ Any R object.}
}
\value{
    Logical determining whether or not argument is Rookable. Not vectorized.
}
\seealso{
\code{\link{Rook}}.
}
\keyword{function}
