% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{print_session_info}
\alias{print_session_info}
\title{print_session_info}
\usage{
print_session_info(s)
}
\arguments{
\item{s}{}
}
\value{
character string with session info
}
\description{
print_session_info
}
