% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_html_from_Rmd}
\alias{render_html_from_Rmd}
\title{render_html_from_Rmd}
\usage{
render_html_from_Rmd(file = NULL, format = NULL, suppress_view = FALSE)
}
\arguments{
\item{file}{}

\item{format}{}
}
\description{
render_html_from_Rmd
}
