\name{Server}
\Rdversion{1.1}
\alias{Server}

\title{Rook Server Object}
\description{
\code{Server} is an object exported by Rook that has no value to the user. It is mainly used by web servers for their convenience. To see an example of how it may be used, see rApache.R in the inst/servers directory.
} 
\keyword{classes}
